Added to the settings in the admin a preview of the callbackURL
