+++ package.json
@@ -24,6 +24,7 @@
     "lodash": "^4.17.21",
     "polished": "^4.1.1",
     "react": "^17.0.2",
+    "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "^17.0.2",
     "react-error-boundary": "^3.1.4",
     "react-router-dom": "^6.6.1",
@@ -41,6 +42,7 @@
     "@types/jest": "^27.4.0",
     "@types/lodash": "^4.14.182",
     "@types/react": "^18.0.26",
+    "@types/react-copy-to-clipboard": "^5.0.4",
     "@types/react-dom": "^18.0.10",
     "@types/showdown": "^2.0.0",
     "@types/styled-components": "^5.1.25",
