--- src/App.css
@@ -1,1 +1,0 @@
-/* App styles */
