--- src/index.css
@@ -1,1 +1,0 @@
-/* Global styles */
