+++ src/pages/Main.tsx
@@ -1,232 +1,24 @@
-import { useEffect, useState, Suspense } from "react";
-import { useDebouncedCallback } from "use-debounce";
-import { Routes, Route, useNavigate } from "react-router-dom";
-import { ErrorBoundary } from "react-error-boundary";
-import { match } from "ts-pattern";
+import { useNavigate } from "react-router-dom";
 import {
-    Context,
-    TargetAction,
-    useDeskproAppClient,
-    useDeskproAppEvents,
+    LoadingSpinner,
     useInitialisedDeskproAppClient,
 } from "@deskpro/app-sdk";
-import { AppElementPayload, ReplyBoxSelection } from "../context/StoreProvider/types";
 import { useStore } from "../context/StoreProvider/hooks";
-import { baseRequest, checkIsAuthService, getIssueCommentUrl } from "../services/github";
-import { useLogout, useUnlinkIssue } from "../hooks";
-import {
-    ticketReplyNotesSelectionStateKey,
-    ticketReplyEmailsSelectionStateKey,
-    registerReplyBoxNotesAdditionsTargetAction,
-    registerReplyBoxEmailsAdditionsTargetAction,
-} from "../utils/replyBox";
-import { LogInPage } from "./LogIn";
-import { HomePage } from "./HomePage";
-import { LinkIssuePage } from "./LinkIssuePage";
-import { ViewIssuePage } from "./ViewIssuePage";
-import { EditIssuePage } from "./EditIssuePage";
-import { AddCommentPage } from "./AddCommentPage";
-import { ErrorBlock, Loading } from "../components/common";
-import { IssueGQL } from "../services/github/types";
+import { checkIsAuthService } from "../services/github";
 
 export const Main = () => {
     const navigate = useNavigate();
-    const [state, dispatch] = useStore();
-    const { client } = useDeskproAppClient();
-    const [loading, setLoading] = useState<boolean>(false);
-    const { logout, isLoading: isLoadingLogout } = useLogout();
-    const { unlinkIssue, isLoading: isLoadingUnlink } = useUnlinkIssue();
-
-    const isLoading = [
-        loading,
-        isLoadingUnlink,
-        isLoadingLogout,
-    ].some((isLoading) => isLoading);
-
-    if (state._error) {
-        // eslint-disable-next-line no-console
-        console.error(`GitHub: ${state._error}`);
-    }
+    const [, dispatch] = useStore();
 
     useInitialisedDeskproAppClient((client) => {
-        setLoading(true);
-
         checkIsAuthService(client)
-            .then((isAuth) => dispatch({ type: "setAuth", isAuth }))
-            .finally(() => setLoading(false));
+            .then((isAuth) => {
+                dispatch({ type: "setAuth", isAuth })
+                navigate(isAuth ? "/home" : "/log_in")
+            })
     });
 
-    useEffect(() => {
-        if (state.isAuth) {
-            navigate("/home");
-        }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [state.isAuth]);
-
-    useInitialisedDeskproAppClient((client) => {
-        registerReplyBoxNotesAdditionsTargetAction(client, state);
-        registerReplyBoxEmailsAdditionsTargetAction(client, state);
-        client.registerTargetAction("githubOnReplyBoxNote", "on_reply_box_note");
-        client.registerTargetAction("githubOnReplyBoxEmail", "on_reply_box_email");
-    }, [state.issues, state.context?.data]);
-
-    const debounceTargetAction = useDebouncedCallback<(a: TargetAction<ReplyBoxSelection[]>) => void>(
-        (action: TargetAction) => {
-            dispatch({ type: "error", error: null });
-
-            match<string>(action.name)
-                .with("linkTicket", () => navigate("/link_issue"))
-                .with("githubOnReplyBoxNote", () => {
-                    const ticketId = action.subject;
-                    const note = action.payload.note;
-
-                    if (!ticketId || !note || !client) {
-                        return;
-                    }
-
-                    if (ticketId !== state.context?.data.ticket.id) {
-                        return;
-                    }
-
-                    client.setBlocking(true);
-                    client.getState<ReplyBoxSelection>(`tickets/${ticketId}/github/notes/*`)
-                        .then((r) => {
-                            const commentUrls = r
-                                .filter(({ data }) => data?.selected)
-                                .map(({ data }) => data?.id)
-                                .map((issueId) => (state.issues ?? []).find(({ id }) => issueId === id.toLowerCase()) as IssueGQL)
-                                .filter((issue) => !!issue)
-                                .map(({ repository, number }) => getIssueCommentUrl(repository.nameWithOwner, number) );
-
-                            return Promise.all(commentUrls.map((commentUrl) => baseRequest(client, {
-                                rawUrl: commentUrl,
-                                method: "POST",
-                                data: {
-                                    body: note,
-                                },
-                            })));
-                        })
-                        .then(() => dispatch({ type: "setIssue", issue: null }))
-                        .finally(() => client.setBlocking(false))
-                })
-                .with("githubOnReplyBoxEmail", () => {
-                    const ticketId = action.subject;
-                    const email = action.payload.email;
-
-                    if (!ticketId || !email || !client) {
-                        return;
-                    }
-
-                    if (ticketId !== state.context?.data.ticket.id) {
-                        return;
-                    }
-
-                    client.setBlocking(true);
-                    client.getState<ReplyBoxSelection>(`tickets/${ticketId}/github/emails/*`)
-                        .then((r) => {
-                            const commentUrls = r
-                                .filter(({ data }) => data?.selected)
-                                .map(({ data }) => data?.id)
-                                .map((issueId) => (state.issues ?? []).find(({ id }) => issueId === id.toLowerCase()) as IssueGQL)
-                                .filter((issue) => !!issue)
-                                .map(({ repository, number }) => getIssueCommentUrl(repository.nameWithOwner, number) );
-
-                            return Promise.all(commentUrls.map((commentUrl) => baseRequest(client, {
-                                rawUrl: commentUrl,
-                                method: "POST",
-                                data: {
-                                    body: email
-                                }
-                            })));
-                        })
-                        .then(() => dispatch({ type: "setIssue", issue: null }))
-                        .finally(() => client.setBlocking(false))
-                })
-                .with("githubReplyBoxNoteAdditions", () => {
-                    (action.payload ?? []).forEach((selection: { id: string; selected: boolean; }) => {
-                        const ticketId = action.subject;
-
-                        if (state.context?.data.ticket.id) {
-                            client?.setState(
-                                ticketReplyNotesSelectionStateKey(ticketId, selection.id),
-                                { id: selection.id, selected: selection.selected }
-                            ).then((result) => {
-                                if (result.isSuccess) {
-                                    registerReplyBoxNotesAdditionsTargetAction(client, state);
-                                } else if (!result.isSuccess && result.errors.length) {
-                                    dispatch({ type: "error", error: result.errors });
-                                }
-                            });
-                        }
-                    })
-                })
-                .with("githubReplyBoxEmailAdditions", () => {
-                    (action.payload ?? []).forEach((selection: { id: string; selected: boolean; }) => {
-                        const ticketId = action.subject;
-
-                        if (state.context?.data.ticket.id) {
-                            client?.setState(
-                                ticketReplyEmailsSelectionStateKey(ticketId, selection.id),
-                                { id: selection.id, selected: selection.selected }
-                            ).then((result) => {
-                                if (result.isSuccess) {
-                                    registerReplyBoxEmailsAdditionsTargetAction(client, state);
-                                } else if (!result.isSuccess && result.errors.length) {
-                                    dispatch({ type: "error", error: result.errors });
-                                }
-                            });
-                        }
-                    })
-                })
-                .run();
-        },
-        500,
+    return (
+        <LoadingSpinner/>
     );
-
-    useDeskproAppEvents({
-        onShow: () => {
-            client && setTimeout(() => client.resize(), 200);
-        },
-        onChange: (context: Context) => {
-            context && dispatch({ type: "loadContext", context });
-        },
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        onElementEvent: (id: string, type: string, payload?: AppElementPayload) => {
-            if (payload?.type === "changePage") {
-                navigate(payload.params);
-            } else if (payload?.type === "logout") {
-                logout();
-            } else if (payload?.type === "unlinkTicket") {
-                unlinkIssue(payload);
-            }
-
-            match(type)
-                .with("home_button", () => dispatch({ type: "setIssue", issue: null }))
-                .otherwise(() => {});
-        },
-        onTargetAction: (a) => debounceTargetAction(a as TargetAction),
-    }, [client, unlinkIssue, logout]);
-
-    return isLoading
-        ? (<Loading />)
-        : (
-            <>
-                {state._error && (<ErrorBlock />)}
-                <Suspense fallback={<Loading />}>
-                    <ErrorBoundary fallbackRender={() => (<ErrorBlock text="An error occurred..." />)}>
-                        <Routes>
-                            <Route path="/home" element={<HomePage />} />
-                            <Route path="/log_in" element={<LogInPage />} />
-                            <Route path="/link_issue" element={<LinkIssuePage />} />
-                            <Route path="/view_issue" element={<ViewIssuePage />} />
-                            <Route path="/edit_issue" element={<EditIssuePage />} />
-                            <Route path="/add_comment" element={<AddCommentPage />} />
-                            <Route index element={<LogInPage />} />
-                        </Routes>
-                    </ErrorBoundary>
-                </Suspense>
-                <br/><br/><br/>
-            </>
-        );
 };
